// gen/ts/src/access/roles.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.8
//   protoc               v5.27.3
// source: access/roles.proto

/* eslint-disable */

export const protobufPackage = "kusinta.iot.v1.access";

export enum Role {
  ROLE_UNSPECIFIED = "ROLE_UNSPECIFIED",
  RESIDENT = "RESIDENT",
  PROPERTY_OWNER = "PROPERTY_OWNER",
  TECHNICIAN = "TECHNICIAN",
  GATEWAY_ADMIN = "GATEWAY_ADMIN",
  UNRECOGNIZED = "UNRECOGNIZED",
}

export function roleFromJSON(object: any): Role {
  switch (object) {
    case 0:
    case "ROLE_UNSPECIFIED":
      return Role.ROLE_UNSPECIFIED;
    case 1:
    case "RESIDENT":
      return Role.RESIDENT;
    case 2:
    case "PROPERTY_OWNER":
      return Role.PROPERTY_OWNER;
    case 3:
    case "TECHNICIAN":
      return Role.TECHNICIAN;
    case 4:
    case "GATEWAY_ADMIN":
      return Role.GATEWAY_ADMIN;
    case -1:
    case "UNRECOGNIZED":
    default:
      return Role.UNRECOGNIZED;
  }
}

export function roleToJSON(object: Role): string {
  switch (object) {
    case Role.ROLE_UNSPECIFIED:
      return "ROLE_UNSPECIFIED";
    case Role.RESIDENT:
      return "RESIDENT";
    case Role.PROPERTY_OWNER:
      return "PROPERTY_OWNER";
    case Role.TECHNICIAN:
      return "TECHNICIAN";
    case Role.GATEWAY_ADMIN:
      return "GATEWAY_ADMIN";
    case Role.UNRECOGNIZED:
    default:
      return "UNRECOGNIZED";
  }
}

export function roleToNumber(object: Role): number {
  switch (object) {
    case Role.ROLE_UNSPECIFIED:
      return 0;
    case Role.RESIDENT:
      return 1;
    case Role.PROPERTY_OWNER:
      return 2;
    case Role.TECHNICIAN:
      return 3;
    case Role.GATEWAY_ADMIN:
      return 4;
    case Role.UNRECOGNIZED:
    default:
      return -1;
  }
}

export enum PermissionAction {
  PERMISSION_ACTION_UNSPECIFIED = "PERMISSION_ACTION_UNSPECIFIED",
  READ = "READ",
  WRITE = "WRITE",
  OBSERVE = "OBSERVE",
  INVOKE = "INVOKE",
  UNRECOGNIZED = "UNRECOGNIZED",
}

export function permissionActionFromJSON(object: any): PermissionAction {
  switch (object) {
    case 0:
    case "PERMISSION_ACTION_UNSPECIFIED":
      return PermissionAction.PERMISSION_ACTION_UNSPECIFIED;
    case 1:
    case "READ":
      return PermissionAction.READ;
    case 2:
    case "WRITE":
      return PermissionAction.WRITE;
    case 3:
    case "OBSERVE":
      return PermissionAction.OBSERVE;
    case 4:
    case "INVOKE":
      return PermissionAction.INVOKE;
    case -1:
    case "UNRECOGNIZED":
    default:
      return PermissionAction.UNRECOGNIZED;
  }
}

export function permissionActionToJSON(object: PermissionAction): string {
  switch (object) {
    case PermissionAction.PERMISSION_ACTION_UNSPECIFIED:
      return "PERMISSION_ACTION_UNSPECIFIED";
    case PermissionAction.READ:
      return "READ";
    case PermissionAction.WRITE:
      return "WRITE";
    case PermissionAction.OBSERVE:
      return "OBSERVE";
    case PermissionAction.INVOKE:
      return "INVOKE";
    case PermissionAction.UNRECOGNIZED:
    default:
      return "UNRECOGNIZED";
  }
}

export function permissionActionToNumber(object: PermissionAction): number {
  switch (object) {
    case PermissionAction.PERMISSION_ACTION_UNSPECIFIED:
      return 0;
    case PermissionAction.READ:
      return 1;
    case PermissionAction.WRITE:
      return 2;
    case PermissionAction.OBSERVE:
      return 3;
    case PermissionAction.INVOKE:
      return 4;
    case PermissionAction.UNRECOGNIZED:
    default:
      return -1;
  }
}